hw-mgmt: scripts: Move shared python code to separate module
Moved logger/timer classes from TC files to python module:

./usr/bin/hw_management_lib.py
